fix: windows tests

+++ mitm/lib/SocketConnectDriver.ts
@@ -5,6 +5,7 @@ import Log from '@secret-agent/commons/Logger';
 import { EventEmitter } from 'events';
 import { createPromise } from '@secret-agent/commons/utils';
 import * as os from 'os';
+import { v1 } from 'uuid';
 
 const { log } = Log(module);
 
@@ -16,6 +17,7 @@ export default class SocketConnectDriver {
   public remoteAddress: string;
   public localAddress: string;
 
+  private isClosing = false;
   private isConnected = false;
   private child: ChildProcess;
   private emitter = new EventEmitter();
@@ -23,8 +25,9 @@ export default class SocketConnectDriver {
   constructor(readonly sessionId: string, readonly connectOpts: IGoTlsSocketConnectOpts) {
     const id = (counter += 1);
     this.socketPath =
-      os.platform() === 'win32' ? `\\\\.\\pipe\\sa-${id}` : `${os.tmpdir()}/sa-mitm-${id}.sock`;
+      os.platform() === 'win32' ? `\\\\.\\pipe\\sa-${v1()}` : `${os.tmpdir()}/sa-mitm-${id}.sock`;
 
+    if (connectOpts.debug === undefined) connectOpts.debug = log.level === 'stats';
     if (connectOpts.isSsl === undefined) connectOpts.isSsl = true;
   }
 
@@ -49,6 +52,8 @@ export default class SocketConnectDriver {
   }
 
   public close() {
+    if (this.isClosing) return;
+    this.isClosing = true;
     this.emitter.emit('close');
     this.cleanupSocket();
     this.closeChild();
@@ -80,11 +85,7 @@ export default class SocketConnectDriver {
     const promise = createPromise(30e3);
     child.on('exit', code => {
       promise.reject(new Error('Socket process exited during connect'));
-      if (this.socket) {
-        this.socket.removeAllListeners();
-        this.socket.end();
-        this.cleanupSocket();
-      }
+      this.cleanupSocket();
     });
 
     child.on('error', error => {
@@ -114,19 +115,21 @@ export default class SocketConnectDriver {
 
   private closeChild() {
     if (this.child.killed) return;
-    this.child.unref();
     try {
       this.child.stdin.write('disconnect');
     } catch (err) {
       // don't log epipes
     }
-    this.child.kill();
+    if (os.platform() !== 'win32') {
+      this.child.kill();
+    }
+    this.child.unref();
   }
 
   private cleanupSocket() {
     if (!this.socket) return;
+    this.socket.end();
     unlink(this.socketPath, () => null);
-    this.socket.removeAllListeners();
     delete this.socket;
   }
